+++ src/whatsapp/controllers/webhook.controller.ts
@@ -48,6 +48,9 @@ export class WebhookController {
         'CONNECTION_UPDATE',
         'CALL',
         'NEW_JWT_TOKEN',
+        'TYPEBOT_START',
+        'TYPEBOT_CHANGE_STATUS',
+        'CHAMA_AI_ACTION',
       ];
     }